events: get_events now is multi-dispatch, returns query-info for catalogues
To cope with the need of having the information when
doing get_events on a catalgoue as to whether an event was
selected because it is assigned to a catalogue or whether
it matched the predicate of a catalogue, get_events become multi-dispatch.

If used with a catalogue it returns a tuple of a list of events and
a list of QueryInformation (only assigned: bool, for the moment). The
indexes of the two lists are matching.

get_event with None or a Predicate has not changed.